Migrate local GitHub actions to TypeScript
