+++ .github/actions/generate-changelog/.gitignore
@@ -1,0 +1,1 @@
+main.js
