+++ .github/actions/publish-github-release/.gitignore
@@ -1,0 +1,1 @@
+main.js
