+++ .github/actions/publish-github-release/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "extends": "../../../tsconfig",
+  "compilerOptions": {
+    "declaration": false,
+    "module": "commonjs",
+    "composite": false,
+    "sourceMap": false,
+    "target": "esnext"
+  },
+  "files": [
+    "./main.ts"
+  ]
+}
