+++ .github/workflows/release.yml
@@ -13,7 +13,7 @@ jobs:
 
       - name: Generate the changelog
         id: changelog
-        uses: cliqz-oss/adblocker/actions/generate-changelog@master
+        uses: cliqz-oss/adblocker/.github/actions/generate-changelog@master
         env:
           GITHUB_AUTH: ${{ secrets.GITHUB_TOKEN }}
 
@@ -34,7 +34,7 @@ jobs:
           force: false
 
       - name: Create a GitHub release
-        uses: cliqz-oss/adblocker/actions/publish-github-release@master
+        uses: cliqz-oss/adblocker/.github/actions/publish-github-release@master
         env:
           GITHUB_AUTH: ${{ secrets.GITHUB_TOKEN }}
         with:
