--- actions/generate-changelog/package.json
@@ -1,10 +1,0 @@
-{
-  "private": true,
-  "name": "update-changelog",
-  "version": "1.0.0",
-  "license": "MPL-2.0",
-  "dependencies": {
-    "@actions/core": "^1.2.1",
-    "lerna-changelog": "^1.0.0"
-  }
-}
