--- actions/publish-github-release/package.json
@@ -1,11 +1,0 @@
-{
-  "private": true,
-  "name": "publish-github-release",
-  "version": "1.0.0",
-  "license": "MPL-2.0",
-  "dependencies": {
-    "@actions/core": "^1.2.1",
-    "@actions/github": "^2.0.1",
-    "lerna-changelog": "^1.0.0"
-  }
-}
