+++ package.json
@@ -49,7 +49,8 @@
     "url": "git+ssh://git@github.com/cliqz-oss/adblocker.git"
   },
   "workspaces": [
-    "./packages/*"
+    "./packages/*",
+    "./.github/actions/*"
   ],
   "scripts": {
     "bootstrap": "npx lerna bootstrap",
